Completely rewrite theme
Theme now renders using Handlebars, rather than plain PHP, and start
wiring up some JavaScript for the page, using WebPack to manage all
the assets. We're going to treat the theme as a link viewer application,
providing contextual data for the links and grouping them by week.

We're using Jaxion to manage the application & functions.php as the
application entry point. We're circumventing the template hierarchy
using Brain\Cortex for routing, and Brain\Context to provide the data
to the templates required. This forces us to split the data querying from
the template rendering, which is nice, but we're still kinda fighting the
standard method of querying data, since we want to group by dates
rather than just outputting a list of links.

Next, we're going to rewrite the CSS to stop using bootstrap and
switch to PostCSS. For now, we'll continue grouping, but that may
change as we build out the application.